[FIX] Category ID is a long
https://github.com/cloudbees/zendesk-java-client/blob/master/src/main/java/org/zendesk/client/v2/model/hc/Category.java#L10
https://developer.zendesk.com/rest_api/docs/help_center/categories#json-format (marked as int but they aren't  - it is the same for others objects)

+++ src/main/java/org/zendesk/client/v2/Zendesk.java
@@ -1586,7 +1586,7 @@ public class Zendesk implements Closeable {
             handleList(Category.class, "categories"));
     }
 
-    public Category getCategory(int id) {
+    public Category getCategory(long id) {
         return complete(submit(req("GET", tmpl("/help_center/categories/{id}.json").set("id", id)),
                 handle(Category.class, "category")));
     }
